+++ sound-mind-studio/tests/test_landing_page.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <QObject>
+
+class LandingPageTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void newProjectButtonEmitsNewProjectRequested();
+    void openProjectButtonEmitsOpenProjectRequested();
+    void setRecentProjectsWithNoPathsShowsNoClickableEntries();
+    void setRecentProjectsCreatesEntriesThatEmitTheirPath();
+    void setRecentProjectsReplacesThePreviousEntries();
+};