+++ sound-mind-studio/tests/test_recent_projects.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <QObject>
+
+class RecentProjectsTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void listIsEmptyForFreshSettings();
+    void addThenListReturnsThatPath();
+    void listFiltersOutPathsThatNoLongerExist();
+    void addMovesAnAlreadyPresentPathToTheFront();
+    void addTrimsToMaxEntries();
+};